requirements.txt
requests==2.30.0
google-api-python-client==2.86.0
google-auth==2.18.0
python-dotenv==1.0.0